🔨 [functional] Move `patch` tests to separate module
